requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.0
requests==2.32.5
PyYAML==6.0.3
urllib3>=2.0.0,<3.0.0
